#CLOUD-328 one more unit test
#VMS-2248 added std::thread replacement (nx::utils::thread). Needed to run on nx1
#VMS-2248 avoiding instanciating std::random_device often since it may throw
#VMS-2248 fixed exception specification
#VMS-2248 removed duplicate randomBuffer implementation. single std::random_device instance is used
#VMS-2248 removed redundant inheritance
#VMS-2248 review fixes
#VMS-2270 fixed: 'find system' multicast packet did not contain cloudSystemId. That resulted in very long cloud system discovery in LAN
#VMS-2391 CloudServerSocket sends listen request after connection to mediator has been restored
#VMS-2391 RetryTimer: added reset to inital state for re-usage
#VMS-2391 added declaration of subscription to stun client tcp reconnect event
#VMS-2391 fixed compile error
#VMS-2410 bpi build uses gcc-4.8.3 once again
#VMS-2410 fixed compile error
#VMS-2410 minimal std::promise/future replacement to address bugs on linux (x86 and arm)
#VMS-2447 fixed unsynchronized variable access
#vms-2340. Transactions refactoring. QnIcon compile fix.
#vms-2354. Low quality archive playback restored.
#vms-2401. Archive camera info file is created on per storage basis rather than per storage pool.
#vms-2401. Archive camera info file is created on per storage basis rather than per storage pool. Merge heads.
#vms. Merge prod_3.0.0 -> dev_3.0.0.
(VMS-2169) - Crash detected
(VMS-2236) - Crash detected
(VMS-2394) - Audio transcoder wont transcode to G711 correctly
(VMS-2403) minor error fixed
(VMS-2403) misprint fixed
(VMS-2403) modify it for version 2.5 + some small fixes
(VMS-2403) new file added to use it from python, some more variables added to current_config
(VMS-2403) typo fixed
(VMS-2432) - servers has the same System name
(VMS-2459) - Server starts too long due to hardware ID detection
2-way audio works with Axis-p1344
Add missing libs to linux deb packages.
Added forceONVIF param for ActiveCam
Better row hover handling in QnTableView
QnCheckableTableView was removed from table_view.h/cpp
Build improvements: do not copy Qt libs on windows
CLOUD-237 review fixes
CLOUD-324 network::cloud::AddressResolver fix, tests and refactoring
CLOUD-324 review fixes
Changed paths in client*.sln to absolute.
Checkbox indicator slightly adjusted to the new style
Code style improvements
Current layout sometimes wasn't highlighted in the resource tree
Customization folder restored
Explicit qt libs copying for arm devices.
Fallback color for QnPaletteColor (to handle in-palette and out-of-palette colors the same way)
Fix compiling after replacing qFill to std::fill.
Fix for hover+selection combination in tree views
Fix for not finding openal on bpi.
Fix linux compile error
Fix rpi build
Fix vcxproj files generation
Fix webadmin customizations
Fixed VariantAnimator warnings at startup
Fixed crash on client closing
Fixed crash on client start
Fixed timeline customization bug (and warning at startup)
Fixed typo in comment
Fixed unintended implicit castability
Header checkboxes consistently aligned with item checkboxes
Improvements for QnCheckableTableView
Internal bugfix
Lighter background for search tab of resource browser
+some code style improvemenets
Linux compile fix
MOBILE-211: Add app version to the side menu.
MOBILE: Android build fixes and transition to qt-5.6.0.
MOBILE: Embed Roboto font into the app.
MOBILE: Move QnObject to Nx module.
MOBILE: New color theme implementation.
MOBILE: Port StackView to labs controls.
MOBILE: Remove unused components.
New style for table views
Non strict parsing of rtsp headers
Now it's possible to specify a set of icon modes to load and corresponding name suffices (like 'hovered', 'selected' etc.)
Palette assignment changed to more specific (QnTreeView --> QnResourceTreeWidget)
QTableView::viewportSizeHint fix in QnTableView
QnCheckableTableView completely rewritten and moved here from table_view.h/cpp
REVIEW-2122
Remove direct zlib dependency.
Remove invalid assert
Resources with empty name were displayed incorrectly in resource tree
Restoration correct vertsion of the file after repository changing
Restored old mobile clients
Review fixes
Simplify Qt libs copying for client.
Simplify Qt libs copying for mediaserver
Some rpi packaging fixes.
Style fixes
Table selection and hover customization
Two-way audio for Axis cameras. first steps
Update openssl version for Android.
VMS-2154 Put all NICs up before calculating HardwareID on Windows
VMS-2229: New style for resource tree view (part 1)
VMS-2229: New style for resource tree view (part 2 - icons)
VMS-2229: New style for resource tree view (part 3 - icons)
VMS-2229: New style for resource tree view (part 4 - new style drawing)
VMS-2229: New style for resource tree view (part 5 - new layout for Search tab)
VMS-2229: New style for resource tree view (part 6 - inplace editor adjustments)
VMS-2268: Owner's name on cloud system tiles
VMS-2268: Owner's name on cloud system tiles. Added 'Your system'
VMS-2268: Owner's name on cloud system tiles. Added full name binding
VMS-2269: Audit trail window restyling (part 1)
VMS-2269: Audit trail window restyling (part 2)
VMS-2293: Create dialog for access rights setup
VMS-2322: There are more than 8 tiles at Cloud login window
VMS-2342 QnTransactionTcpProcessor temporary socket mode fix
VMS-2362: NX client crash on toolbar logout from cloud at welcome screen
VMS-2384: Create sound-visualizer on mic button instead of fixed animation
VMS-2387 - compile error fix
VMS-2387: Two-Way Audio UI
VMS-2387: Two-Way Audio UI - Hardcode removed
VMS-2391 nx::stun::AsyncClient reconnect tuning
VMS-2391 review fixes
VMS-2392
VMS-2413
VMS-2419
VMS-2419 - desktop camera doesn't found sometimes
VMS-2420 fix possible race in AxisResource
VMS-2420 replace deprecated Qt Algorithms
VMS-2421: GUID of client (for statistics) is changed from hardware-id (takes long time to acquire) to 'pcUuid' generated randomly at first client run and stored in the settings
VMS-326 SSL refactoring, tests and review fixes
VMS-326 SSL sockets tests and fixes
VMS-326 cloud_connect_test_util --ssl support (tested) and mediaserver cloud ssl
VMS-326 review fix
add missing files
added ignoreONVIF param
address review commments
auto value for clientId parameter
avoiding std::async since it is buggy in GNU STL
avoiding std::thread since it is buggy in GNU STL
chmod shell scripts
compile error fixed
cox-added missing resources for old Android client
fixed compile error
fixed compile error in unit tests
fixed compile error on gcc 4.8
fixed compile errors and warningsa
fixed compile warning
fixed compile warnings
fixed compile warnings/errors
fixed compile/link errors
fixed nx1 build
fixing unit tests on arm
future.h moved to common_libs/nx_utils/src/nx/utils/std/
getting rid of std::future/promise in unit tests
hasCameraCapabilities fix
hungarian translation.
minor compile fix
no std::async anymore
removed unused config parameter
rolled back socket test to concurrent recv/send operations
some review fixes
static_unique_ptr_cast moved to <nx/utils/type_utils.h>
tomporarry cloud_db_ut comment for -Dut

<==== Mercurial data ====>
Branch: dev_3.0.0_web

+++ mediaserver_core/src/plugins/resource/onvif/gsoap_async_call_wrapper.h
@@ -251,7 +251,7 @@ private:
             return m_resultHandler( SOAP_FAULT );
         }
 
-        NX_ASSERT( bytesSent == m_serializedRequest.size() );
+        NX_ASSERT( bytesSent == (size_t)m_serializedRequest.size() );
         m_state = receivingResponse;
 
         m_responseBuffer.reserve( READ_BUF_SIZE );
